[WIP] Simplify //expand and //contract
This commit factors a bunch of logic out of expand and contract so that
instead of a big switch statement selecting whether to call
set_first_position or set_second_position, the Player instead has
methods to set a position by an argument.

This also adds a mechanism to compare block positions with respect to a
direction, and to shift blocks with respect to a direction symbolically;
i.e. to take an existing `pos` and construct a new position 100 blocks
North of it, with this patch one would write:
```
    pos + BlockPos::from_direction(BlockFacing::North) * 100
```

This is WIP because there are still directionality bugs I haven't
addressed, and this is generally very poorly tested, but I'd like to
know if the project likes this direction at all before I spend time
fixing annoying sign bugs. The change does work at least somewhat now,
just you might expand or contract in the wrong direction or move the
wrong point.